% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMUC_2020.R
\docType{data}
\name{IMC_2020}
\alias{IMC_2020}
\title{Base de índice de marginación a nivel colonia, 2020}
\format{
A data frame with 74226 observations on the following 25 variables.
}
\usage{
IMC_2020
}
\description{
Base de datos del índice de marginación a nivel colonia urbana, 2020. Esta base contiene a la población total, indicadores
sociodemográficos, índice y grado de marginación.
\url{https://datos.gob.mx/busca/dataset/indice-de-marginacion-carencias-poblacionales-por-localidad-municipio-y-entidad}
}
\section{Variables}{


\itemize{
\item{\code{CVE_COL}}{- Clave de colonia}
\item{\code{ID_COL}}{- ID de la colonia}
\item{\code{ID_INE_2020}}{- ID de la colonia (INE)}
\item{\code{NOM_COLONIA}}{- Nombre de la colonia}
\item{\code{CP}}{- Código Postal}
\item{\code{CLASIF}}{- Clasificación de colonias}
\item{\code{CVE_ENT}}{- Clave de la entidad}
\item{\code{NOM_ENT}}{- Nombre de la entidad}
\item{\code{CVE_MUN}}{- Clave del municipio}
\item{\code{NOM_ENT}}{- Nombre del municipio}
\item{\code{POB_TOT}}{- Población total 2020}
\item{\code{P6A14NAE}}{- Porcentaje de población de 6 a 14 años que no asiste a la escuela}
\item{\code{SBASC}}{- Porcentaje de población de 15 años o mas sin educación básica}
\item{\code{PSDSS}}{- Porcentaje de población sin derechohabiencia a los servicios de salud}
\item{\code{OVSDE}}{- Porcentaje de ocupantes en viviendas particulares habitadas sin drenaje ni excusado}
\item{\code{OVSEE}}{- Porcentaje de ocupantes en viviendas particulares habitadas sin agua entubada}
\item{\code{OVSAE}}{- Porcentaje de ocupantes en viviendas particulares habitadas con piso de tierra}
\item{\code{OVPT}}{- Porcentaje de ocupantes en viviendas particulares habitadas con piso de tierra}
\item{\code{OVHAC}}{- Porcentaje de ocupantes en viviendas particulares habitadas con hacinamiento}
\item{\code{OVSREF}}{-Porcentaje de ocupantes en viviendas particulares habitadas sin refrigerador}
\item{\code{OVSINT}}{-Porcentaje de ocupantes en viviendas particulares habitadas sin internet}
\item{\code{OVSCEL}}{-Porcentaje de ocupantes en viviendas particulares habitadas sin celular}
\item{\code{IM_2020}}{- Índice de marginación, 2020}
\item{\code{GM_2020}}{- Grado de marginación, 2020}
\item{\code{IMN_2020}}{- Índice de marginación normalizado, 2020}
}
}

\examples{
\dontrun{head(IMC_2020)}
}
\keyword{datasets}
